requirements.txt
numpy
pygame
matplotlib
torch
torchvision
Pillow
opencv-python